requirements.txt
numpy>=1.24
torch>=2.1
PyYAML>=6.0
networkx>=3.2
scipy>=1.11
pytest>=8.0
